Add some exceptions and user error msgs to game deletion

+++ Provenance/Game Library/Database/RomDatabase.swift
@@ -292,6 +292,10 @@ public extension RomDatabase {
     }
 }
 
+public enum RomDeletionError : Error {
+	case relatedFiledDeletionError
+}
+
 // MARK: - Update
 public extension RomDatabase {
     @objc
@@ -350,7 +354,7 @@ public extension RomDatabase {
 		}
 	}
 
-	func delete(game: PVGame) {
+	func delete(game: PVGame) throws {
 		let romURL = PVEmulatorConfiguration.path(forGame: game)
 
 		if !game.customArtworkURL.isEmpty {
@@ -358,27 +362,33 @@ public extension RomDatabase {
 				try PVMediaCache.deleteImage(forKey: game.customArtworkURL)
 			} catch {
 				ELOG("Failed to delete image " + game.customArtworkURL)
+				// Don't throw, not a big deal
 			}
 		}
 
 		let savesPath = PVEmulatorConfiguration.saveStatePath(forGame: game)
-		do {
-			try FileManager.default.removeItem(at: savesPath)
-		} catch {
-			WLOG("Unable to delete save states at path: " + savesPath.path + "because: " + error.localizedDescription)
+		if FileManager.default.fileExists(atPath: savesPath.path) {
+			do {
+				try FileManager.default.removeItem(at: savesPath)
+			} catch {
+				ELOG("Unable to delete save states at path: " + savesPath.path + "because: " + error.localizedDescription)
+			}
 		}
 
 		let batteryPath = PVEmulatorConfiguration.batterySavesPath(forGame: game)
-		do {
-			try FileManager.default.removeItem(at: batteryPath)
-		} catch {
-			WLOG("Unable to delete battery states at path: \(batteryPath.path) because: \(error.localizedDescription)")
+		if FileManager.default.fileExists(atPath: batteryPath.path) {
+			do {
+				try FileManager.default.removeItem(at: batteryPath)
+			} catch {
+				ELOG("Unable to delete battery states at path: \(batteryPath.path) because: \(error.localizedDescription)")
+			}
 		}
 
 		do {
 			try FileManager.default.removeItem(at: romURL)
 		} catch {
-			WLOG("Unable to delete rom at path: \(romURL.path) because: \(error.localizedDescription)")
+			ELOG("Unable to delete rom at path: \(romURL.path) because: \(error.localizedDescription)")
+			throw error
 		}
 
 		// Delete from Spotlight search
@@ -392,19 +402,18 @@ public extension RomDatabase {
 		game.recentPlays.forEach { try? $0.delete() }
 		game.screenShots.forEach { try? $0.delete() }
 
-		deleteRelatedFilesGame(game)
-		try? game.delete()
+		try deleteRelatedFilesGame(game)
+		try game.delete()
 	}
 
-	func deleteRelatedFilesGame(_ game: PVGame) {
-
-		game.relatedFiles.forEach {
-			try? FileManager.default.removeItem(at: $0.url )
-		}
-		
+	func deleteRelatedFilesGame(_ game: PVGame) throws {
 		guard let system = game.system else {
 			ELOG("Game \(game.title) belongs to an unknown system \(game.systemIdentifier)")
-			return
+			throw RomDeletionError.relatedFiledDeletionError
+		}
+
+		try game.relatedFiles.forEach {
+			try FileManager.default.removeItem(at: $0.url )
 		}
 
 		let romDirectory = system.romsDirectory
@@ -423,12 +432,13 @@ public extension RomDatabase {
 			return filename.contains(relatedFileName)
 		}
 
-		matchingFiles.forEach {
+		try matchingFiles.forEach {
 			let file = romDirectory.appendingPathComponent( $0.lastPathComponent, isDirectory: false)
 			do {
 				try FileManager.default.removeItem(at: file)
 			} catch {
 				ELOG("Failed to remove item \(file.path).\n \(error.localizedDescription)")
+				throw error
 			}
 		}
 	}
